Enabling gloo CI (#582)
* new openmpi in conda

* adding new ci file

* minor fix and adding ucc

* debugging

* debugging

* debugging

* minor changes

* removing default nullptr MPI_Comm

* fixing cython

* temp remove windows c++ tests

* attempting to fix

* attempting to fix

* changing dist_sort_test eval func

* fixing problems

* adding windows CI

* minor change

* minor changes

+++ python/pycylon/test/test_series.py
@@ -41,7 +41,7 @@ def test_series_with_list():
 
 
 def test_series_with_numpy():
-    ld = np.array([1, 2, 3, 4])
+    ld = np.array([1, 2, 3, 4], dtype=np.int64)
     id = 's1'
     dtype = cn.int64()
     s = Series(id, ld)
